requirements.txt
django==5.0.8
djangorestframework~=3.15.2
django-debug-toolbar==4.4.6
